add cdi
