+++ gradle.properties
@@ -15,13 +15,13 @@ lombokPluginVersion=6.4.1
 quarkusVersion=2.9.2.Final
 junitVersion=5.8.1
 jakartaValidationApiVersion=2.0.2
-guavaVersion=30.1.1-jre
+guavaVersion=31.1-jre
 slf4jVersion=1.7.30
 hibernateVersion=5.6.1.Final
 hamcrestVersion=1.3
 mockitoVersion=4.1.0
 jacksonVersion=2.12.3
-graphqlJavaVersion=17.3
+graphqlJavaVersion=18.1
 validationApiVersion=2.0.1.Final
 lombokVersion=1.18.22
 cdiApiVersion=2.0.2
